refactor: split js.ts into 5 focused modules (1687 → 444 lines)
Extract pulse layout, suggest improvements, build-from-goal wizard,
and template palette autocomplete into separate *.js.ts files.
layout.ts concatenates all modules. No behavior change.
