+++ packages/dashboard/src/views/js.ts
@@ -1,9 +1,13 @@
 /**
- * Inline JS. Tiny vanilla script for:
- *   1. The community-shell confirm modal (open/close, checkbox-gated submit)
- *   2. Auto-poll on /runs/:id when the run is still in-progress
+ * Inline JS — core dashboard utilities.
  *
- * Inlined so there's no second HTTP round-trip for ~2KB of logic.
+ * Large feature modules (pulse layout, suggest improvements, build-from-goal,
+ * template palette) have been extracted to separate *.js.ts files and are
+ * concatenated in layout.ts.
+ *
+ * What remains here: community-shell modal, tool-picker, node-type toggle,
+ * ESC handler, confirm-before-submit, node-vars filter, run status poll,
+ * run-now modal, LLM provider/model sync, secret save confirmation.
  */
 export const DASHBOARD_JS = `
 (function () {
@@ -195,327 +199,6 @@ export const DASHBOARD_JS = `
     }
   });
 
-  // ── Template palette autocomplete ──────────────────────────────────
-  // Autocomplete for $ (shell env vars) and {{
-  // (claude-code template refs) in node command / prompt textareas.
-  // Triggered by typing the first char of either syntax; filters as the
-  // user types; Up/Down + Enter to insert; Esc to close.
-  (function () {
-    var palette = null;
-    var activeTextarea = null;
-    // { mode: 'shell' | 'claude', triggerStart: number, query: string, items: [...], selectedIndex: number }
-    var state = null;
-
-    function getSuggestions(textarea) {
-      var sourceId = textarea.getAttribute('data-palette-source');
-      if (!sourceId) return null;
-      var el = document.getElementById(sourceId);
-      if (!el) return null;
-      try { return JSON.parse(el.textContent); }
-      catch (e) { return null; }
-    }
-
-    function buildItems(mode, sugg) {
-      var items = [];
-      if (mode === 'shell') {
-        for (var i = 0; i < sugg.upstreams.length; i++) {
-          var id = sugg.upstreams[i];
-          var envName = 'UPSTREAM_' + id.toUpperCase().replace(/-/g, '_') + '_RESULT';
-          items.push({
-            insert: '$' + envName,
-            label: '$' + envName,
-            hint: 'stdout of upstream "' + id + '"',
-            group: 'upstream',
-          });
-        }
-        for (var j = 0; j < sugg.inputs.length; j++) {
-          items.push({
-            insert: '$' + sugg.inputs[j],
-            label: '$' + sugg.inputs[j],
-            hint: 'agent input',
-            group: 'input',
-          });
-        }
-        for (var k = 0; k < (sugg.vars || []).length; k++) {
-          items.push({
-            insert: '$' + sugg.vars[k],
-            label: '$' + sugg.vars[k],
-            hint: 'global variable',
-            group: 'var',
-          });
-        }
-        for (var l = 0; l < sugg.secrets.length; l++) {
-          items.push({
-            insert: '$' + sugg.secrets[l],
-            label: '$' + sugg.secrets[l],
-            hint: 'node secret',
-            group: 'secret',
-          });
-        }
-      } else {
-        for (var m = 0; m < sugg.upstreams.length; m++) {
-          var uid = sugg.upstreams[m];
-          items.push({
-            insert: '{{upstream.' + uid + '.result}}',
-            label: '{{upstream.' + uid + '.result}}',
-            hint: 'stdout of upstream "' + uid + '"',
-            group: 'upstream',
-          });
-        }
-        for (var n = 0; n < sugg.inputs.length; n++) {
-          items.push({
-            insert: '{{inputs.' + sugg.inputs[n] + '}}',
-            label: '{{inputs.' + sugg.inputs[n] + '}}',
-            hint: 'agent input',
-            group: 'input',
-          });
-        }
-        for (var p = 0; p < (sugg.vars || []).length; p++) {
-          items.push({
-            insert: '{{vars.' + sugg.vars[p] + '}}',
-            label: '{{vars.' + sugg.vars[p] + '}}',
-            hint: 'global variable',
-            group: 'var',
-          });
-        }
-      }
-      return items;
-    }
-
-    function detectTrigger(textarea) {
-      var v = textarea.value;
-      var pos = textarea.selectionStart;
-      // Walk backwards from the cursor looking for $ or {{. Bail out on
-      // whitespace or on another trigger char — keeps the palette scoped
-      // to the current "word" the user is typing.
-      for (var i = pos - 1; i >= 0 && i >= pos - 64; i--) {
-        var c = v.charAt(i);
-        if (c === '$') {
-          return { mode: 'shell', triggerStart: i, query: v.slice(i + 1, pos) };
-        }
-        if (c === '{' && i > 0 && v.charAt(i - 1) === '{') {
-          return { mode: 'claude', triggerStart: i - 1, query: v.slice(i + 1, pos) };
-        }
-        if (/\\s/.test(c)) return null;
-      }
-      return null;
-    }
-
-    function filterItems(items, query) {
-      if (!query) return items.slice(0, 10);
-      var q = query.toLowerCase();
-      var scored = [];
-      for (var i = 0; i < items.length; i++) {
-        var label = items[i].label.toLowerCase();
-        var idx = label.indexOf(q);
-        if (idx >= 0) scored.push({ item: items[i], score: idx });
-      }
-      scored.sort(function (a, b) { return a.score - b.score; });
-      return scored.slice(0, 10).map(function (s) { return s.item; });
-    }
-
-    function ensurePalette() {
-      if (palette) return palette;
-      palette = document.createElement('div');
-      palette.className = 'template-palette';
-      palette.setAttribute('role', 'listbox');
-      document.body.appendChild(palette);
-      return palette;
-    }
-
-    function closePalette() {
-      if (palette) palette.style.display = 'none';
-      activeTextarea = null;
-      state = null;
-    }
-
-    function positionPalette(textarea) {
-      ensurePalette();
-      var rect = textarea.getBoundingClientRect();
-      // Estimate cursor Y position within the textarea using a mirror div.
-      // This places the palette near the line being edited, not at the
-      // bottom of the entire textarea.
-      var cursorTop = 0;
-      try {
-        var mirror = document.createElement('div');
-        var cs = window.getComputedStyle(textarea);
-        mirror.style.cssText = 'position:absolute;visibility:hidden;white-space:pre-wrap;word-wrap:break-word;overflow:hidden;';
-        mirror.style.width = cs.width;
-        mirror.style.font = cs.font;
-        mirror.style.padding = cs.padding;
-        mirror.style.border = cs.border;
-        mirror.style.lineHeight = cs.lineHeight;
-        mirror.style.letterSpacing = cs.letterSpacing;
-        var textBefore = textarea.value.substring(0, textarea.selectionStart);
-        mirror.textContent = textBefore;
-        var marker = document.createElement('span');
-        marker.textContent = '|';
-        mirror.appendChild(marker);
-        document.body.appendChild(mirror);
-        cursorTop = marker.offsetTop - textarea.scrollTop;
-        document.body.removeChild(mirror);
-      } catch (e) {
-        cursorTop = 0;
-      }
-      // Position palette just below the cursor line, clamped within the textarea bounds.
-      var lineHeight = parseInt(window.getComputedStyle(textarea).lineHeight, 10) || 20;
-      var top = rect.top + cursorTop + lineHeight + 4;
-      // Don't let it go above the textarea or below the viewport.
-      top = Math.max(top, rect.top + lineHeight);
-      top = Math.min(top, rect.bottom + 4);
-      palette.style.left = Math.round(rect.left + window.scrollX) + 'px';
-      palette.style.top = Math.round(top + window.scrollY) + 'px';
-      palette.style.minWidth = Math.round(Math.min(rect.width, 420)) + 'px';
-    }
-
-    var GROUP_LABELS = {
-      upstream: 'Upstream outputs',
-      input: 'Agent inputs',
-      var: 'Global variables',
-      secret: 'Secrets',
-    };
-
-    function renderPalette(items, selectedIndex) {
-      ensurePalette();
-      palette.innerHTML = '';
-      if (items.length === 0) {
-        var empty = document.createElement('div');
-        empty.className = 'template-palette__empty';
-        empty.textContent = 'No matches';
-        palette.appendChild(empty);
-      } else {
-        var lastGroup = null;
-        for (var i = 0; i < items.length; i++) {
-          if (items[i].group !== lastGroup) {
-            lastGroup = items[i].group;
-            var sep = document.createElement('div');
-            sep.className = 'template-palette__group';
-            sep.textContent = GROUP_LABELS[lastGroup] || lastGroup;
-            palette.appendChild(sep);
-          }
-
-          var row = document.createElement('div');
-          row.className = 'template-palette__item' + (i === selectedIndex ? ' is-selected' : '');
-          row.setAttribute('data-index', String(i));
-          row.setAttribute('role', 'option');
-
-          var label = document.createElement('span');
-          label.className = 'template-palette__label mono';
-          label.textContent = items[i].label;
-          row.appendChild(label);
-
-          var hint = document.createElement('span');
-          hint.className = 'template-palette__hint';
-          hint.textContent = items[i].hint;
-          row.appendChild(hint);
-
-          palette.appendChild(row);
-        }
-      }
-      palette.style.display = 'block';
-    }
-
-    function insertChoice(item) {
-      if (!activeTextarea || !state) return;
-      var t = activeTextarea;
-      var before = t.value.slice(0, state.triggerStart);
-      var after = t.value.slice(t.selectionStart);
-      t.value = before + item.insert + after;
-      var newPos = (before + item.insert).length;
-      t.selectionStart = t.selectionEnd = newPos;
-      closePalette();
-      t.focus();
-      // Dispatch an input event so any listeners (form validation, etc.)
-      // see the synthetic insertion.
-      t.dispatchEvent(new Event('input', { bubbles: true }));
-    }
-
-    function updateFromTextarea(textarea) {
-      var trig = detectTrigger(textarea);
-      if (!trig) { closePalette(); return; }
-      var sugg = getSuggestions(textarea);
-      if (!sugg) { closePalette(); return; }
-      var mode = textarea.getAttribute('data-template-palette');
-      // For tool-input fields (mode="both"), show whichever syntax the
-      // user is typing. For dedicated Command/Prompt textareas, enforce
-      // the single syntax. "both" is set on dynamically generated tool
-      // input fields where the user may use either syntax.
-      if (mode !== 'both') {
-        if (mode === 'claude' && trig.mode === 'shell') { closePalette(); return; }
-        if (mode === 'shell' && trig.mode === 'claude') { closePalette(); return; }
-      }
-
-      var all = buildItems(trig.mode, sugg);
-      var filtered = filterItems(all, trig.query);
-
-      activeTextarea = textarea;
-      state = {
-        mode: trig.mode,
-        triggerStart: trig.triggerStart,
-        query: trig.query,
-        items: filtered,
-        selectedIndex: 0,
-      };
-      positionPalette(textarea);
-      renderPalette(filtered, 0);
-    }
-
-    document.addEventListener('input', function (e) {
-      var t = e.target;
-      if (t && t.matches && t.matches('textarea[data-template-palette]')) {
-        updateFromTextarea(t);
-      }
-    });
-
-    document.addEventListener('keydown', function (e) {
-      if (!state || !activeTextarea) return;
-      if (e.target !== activeTextarea) return;
-      if (e.key === 'Escape') {
-        e.preventDefault();
-        closePalette();
-        return;
-      }
-      if (state.items.length === 0) return;
-      if (e.key === 'ArrowDown') {
-        e.preventDefault();
-        state.selectedIndex = (state.selectedIndex + 1) % state.items.length;
-        renderPalette(state.items, state.selectedIndex);
-      } else if (e.key === 'ArrowUp') {
-        e.preventDefault();
-        state.selectedIndex = (state.selectedIndex - 1 + state.items.length) % state.items.length;
-        renderPalette(state.items, state.selectedIndex);
-      } else if (e.key === 'Enter' || e.key === 'Tab') {
-        // Swallow Enter/Tab only when the palette has a concrete pick —
-        // otherwise let the textarea behave normally.
-        e.preventDefault();
-        insertChoice(state.items[state.selectedIndex]);
-      }
-    });
-
-    document.addEventListener('click', function (e) {
-      if (!palette) return;
-      var row = e.target && e.target.closest && e.target.closest('.template-palette__item');
-      if (row && state) {
-        var idx = parseInt(row.getAttribute('data-index'), 10);
-        if (!isNaN(idx) && state.items[idx]) {
-          insertChoice(state.items[idx]);
-          return;
-        }
-      }
-      // Click outside the palette closes it.
-      if (activeTextarea && e.target !== activeTextarea && !(palette && palette.contains(e.target))) {
-        closePalette();
-      }
-    });
-
-    document.addEventListener('blur', function (e) {
-      if (e.target === activeTextarea) {
-        // Delay close so a click on a palette row is still registered.
-        setTimeout(closePalette, 120);
-      }
-    }, true);
-  })();
-
   // ── Node vars filter ───────────────────────────────────────────────
   // Live filter by name or value on resolved variables
   (function () {
@@ -632,764 +315,6 @@ export const DASHBOARD_JS = `
     runModal.addEventListener('click', function (e) {
       if (e.target === runModal) runModal.classList.remove('is-open');
       if (e.target && e.target.getAttribute && e.target.getAttribute('data-close-modal')) runModal.classList.remove('is-open');
-    });
-  })();
-
-  // ── Suggest improvements modal ─────────────────────────────────────
-  // Modal with progress feedback, cancel, colored diff.
-  (function () {
-    var btn = document.getElementById('suggest-btn');
-    var modal = document.getElementById('suggest-modal');
-    var content = document.getElementById('suggest-modal-content');
-    if (!btn || !modal || !content) return;
-    var agentId = btn.getAttribute('data-agent-id');
-    var PHASES = [
-      [0,'Sending YAML to Claude...'],[5,'Reading the DAG structure...'],
-      [15,'Analyzing cross-node data flow...'],[30,'Generating suggestions...'],
-      [60,'Still working (complex agent)...'],[90,'Almost there...']
-    ];
-
-    function esc(s) { var d = document.createElement('div'); d.textContent = s; return d.innerHTML; }
-    function closeModal() { modal.classList.remove('is-open'); }
-
-    // Lightweight markdown to HTML for analysis output.
-    function renderMd(text) {
-      var h = esc(text);
-      h = h.replace(/\\*\\*(.+?)\\*\\*/g, '<strong>' + '$' + '1</strong>');
-      h = h.replace(/\\*(.+?)\\*/g, '<em>' + '$' + '1</em>');
-      h = h.replace(/^- (.+)$/gm, function(m,p1) { return '<li style="margin-left:var(--space-4);list-style:disc;">' + p1 + '</li>'; });
-      h = h.replace(/\\n{2,}/g, '<br><br>');
-      h = h.replace(/\\n/g, '<br>');
-      return h;
-    }
-
-    function coloredDiff(oldT, newT) {
-      var oL = oldT.split('\\n'), nL = newT.split('\\n');
-      var oS = {}, nS = {};
-      for (var i = 0; i < oL.length; i++) oS[oL[i].trim()] = true;
-      for (var j = 0; j < nL.length; j++) nS[nL[j].trim()] = true;
-      var DEL = 'background:rgba(255,0,0,0.08);color:#cf222e;';
-      var ADD = 'background:rgba(0,180,0,0.08);color:#1a7f37;';
-      var P = 'font-size:var(--font-size-xs);background:var(--color-surface-raised);border:1px solid var(--color-border);border-radius:var(--radius-sm);padding:var(--space-3);max-height:280px;overflow-y:auto;white-space:pre-wrap;word-break:break-all;margin:0;';
-      var h = '<div style="display:grid;grid-template-columns:1fr 1fr;gap:var(--space-3);margin-bottom:var(--space-3);">';
-      h += '<div><p class="dim" style="font-size:var(--font-size-xs);margin:0 0 var(--space-1);">Current</p><pre style="' + P + '">';
-      for (var a = 0; a < oL.length; a++) h += '<span style="display:block;' + (nS[oL[a].trim()] ? '' : DEL) + '">' + esc(oL[a]) + '</span>';
-      h += '</pre></div><div><p class="dim" style="font-size:var(--font-size-xs);margin:0 0 var(--space-1);">Suggested</p><pre style="' + P + '">';
-      for (var b = 0; b < nL.length; b++) h += '<span style="display:block;' + (oS[nL[b].trim()] ? '' : ADD) + '">' + esc(nL[b]) + '</span>';
-      h += '</pre></div></div>';
-      return h;
-    }
-
-    function renderResult(data) {
-      var bc = data.classification === 'NO_IMPROVEMENTS' ? 'badge--ok' : data.classification === 'REWRITE' ? 'badge--err' : 'badge--warn';
-      var bl = data.classification === 'NO_IMPROVEMENTS' ? 'No improvements needed' : data.classification === 'REWRITE' ? 'Recommend rewrite' : 'Suggested improvements';
-      var h = '<div style="display:flex;align-items:center;gap:var(--space-3);margin-bottom:var(--space-3);"><span class="badge ' + bc + '">' + esc(bl) + '</span></div>';
-      if (data.summary) h += '<p style="font-weight:var(--weight-medium);margin:0 0 var(--space-3);">' + renderMd(data.summary) + '</p>';
-      if (data.details) h += '<div style="font-size:var(--font-size-sm);line-height:1.6;margin:0 0 var(--space-3);color:var(--color-text-muted);max-height:250px;overflow-y:auto;">' + renderMd(data.details) + '</div>';
-      if (data.yaml && data.currentYaml) {
-        h += coloredDiff(data.currentYaml, data.yaml);
-      } else if (data.yaml) {
-        h += '<pre style="font-size:var(--font-size-xs);background:var(--color-surface-raised);border:1px solid var(--color-border);border-radius:var(--radius-sm);padding:var(--space-3);margin-bottom:var(--space-3);max-height:300px;overflow-y:auto;white-space:pre-wrap;">' + esc(data.yaml) + '</pre>';
-      }
-      if (data.yamlError) {
-        h += '<div class="flash flash--error" style="margin-bottom:var(--space-3);font-size:var(--font-size-xs);">' +
-          '<strong>Suggested YAML has validation errors:</strong> ' + esc(data.yamlError) +
-          '<br>Click "Edit YAML" to fix manually.</div>';
-      }
-      h += '<div style="display:flex;gap:var(--space-2);justify-content:flex-end;flex-wrap:wrap;">';
-      if (data.yaml && !data.yamlError) {
-        h += '<button type="button" class="btn btn--primary btn--sm" id="sg-apply-now">Apply now</button>';
-        h += '<button type="button" class="btn btn--ghost btn--sm" id="sg-review">Review first</button>';
-      } else if (data.yaml) {
-        h += '<button type="button" class="btn btn--primary btn--sm" id="sg-review">Edit YAML to fix</button>';
-      }
-      h += '<button type="button" class="btn btn--ghost btn--sm" id="sg-dismiss">Dismiss</button></div>';
-      content.innerHTML = h;
-      // "Apply now" — save the YAML directly without opening the editor.
-      var an = document.getElementById('sg-apply-now');
-      if (an && data.yaml) an.addEventListener('click', function () {
-        var f = document.createElement('form'); f.method = 'POST';
-        f.action = '/agents/' + encodeURIComponent(agentId) + '/yaml';
-        var t = document.createElement('textarea'); t.name = 'yaml'; t.value = data.yaml; t.style.display = 'none';
-        f.appendChild(t); document.body.appendChild(f); f.submit();
-      });
-      // "Review first" — open the YAML editor pre-filled.
-      var rv = document.getElementById('sg-review');
-      if (rv && data.yaml) rv.addEventListener('click', function () {
-        var f = document.createElement('form'); f.method = 'POST';
-        f.action = '/agents/' + encodeURIComponent(agentId) + '/yaml';
-        var t = document.createElement('textarea'); t.name = 'prefillYaml'; t.value = data.yaml; t.style.display = 'none';
-        f.appendChild(t); document.body.appendChild(f); f.submit();
-      });
-      var db = document.getElementById('sg-dismiss');
-      if (db) db.addEventListener('click', closeModal);
-    }
-
-    btn.addEventListener('click', function () {
-      modal.classList.add('is-open');
-
-      // Step 1: Show a prompt form so the user can optionally describe what to focus on.
-      content.innerHTML =
-        '<div style="padding:var(--space-4);">' +
-          '<h3 style="margin:0 0 var(--space-2);">Suggest improvements for ' + esc(agentId) + '</h3>' +
-          '<p class="dim" style="font-size:var(--font-size-xs);margin:0 0 var(--space-3);">Optionally describe what you want the analysis to focus on. Leave blank for a general review.</p>' +
-          '<textarea id="sg-focus" rows="3" placeholder="e.g. &quot;Are there missing error handlers?&quot; or &quot;The last run timed out, what can I improve?&quot;" style="width:100%;padding:var(--space-2) var(--space-3);border:1px solid var(--color-border-strong);border-radius:var(--radius-sm);font-size:var(--font-size-sm);font-family:var(--font-mono);resize:vertical;background:var(--color-surface-raised);color:var(--color-text);"></textarea>' +
-          '<div style="display:flex;gap:var(--space-2);justify-content:flex-end;margin-top:var(--space-3);">' +
-            '<button type="button" class="btn btn--ghost btn--sm" id="sg-prompt-cancel">Cancel</button>' +
-            '<button type="button" class="btn btn--primary btn--sm" id="sg-prompt-go">Analyze</button>' +
-          '</div>' +
-        '</div>';
-      document.getElementById('sg-prompt-cancel').addEventListener('click', closeModal);
-      document.getElementById('sg-focus').focus();
-
-      document.getElementById('sg-prompt-go').addEventListener('click', function () {
-      var focusText = (document.getElementById('sg-focus').value || '').trim();
-
-      // Step 2: Switch to progress view and start analysis.
-      var t0 = Date.now();
-      var cancelled = false;
-      var pollTimer = null;
-      var tickTimer = null;
-
-      function showProgress(phaseMsg) {
-        var s = Math.round((Date.now() - t0) / 1000);
-        content.innerHTML =
-          '<div style="padding:var(--space-4);">' +
-          '<div style="display:flex;align-items:center;gap:var(--space-3);margin-bottom:var(--space-4);">' +
-            '<div class="spinner"></div>' +
-            '<div style="flex:1;"><div style="font-weight:var(--weight-medium);">Analyzing ' + esc(agentId) + '</div>' +
-            '<div class="dim" style="font-size:var(--font-size-xs);" id="sg-phase">' + esc(phaseMsg) + '</div></div>' +
-            '<div style="font-family:var(--font-mono);font-size:var(--font-size-lg);color:var(--color-text-muted);min-width:3rem;text-align:right;" id="sg-timer">' + s + 's</div>' +
-          '</div>' +
-          '<div style="height:4px;background:var(--color-border);border-radius:2px;overflow:hidden;margin-bottom:var(--space-3);">' +
-            '<div style="height:100%;background:var(--color-primary);border-radius:2px;width:' + Math.min(s/120*100, 98) + '%;transition:width 1s linear;"></div>' +
-          '</div>' +
-          '<div style="text-align:right;"><button type="button" class="btn btn--ghost btn--sm" id="sg-cancel">Cancel</button></div>' +
-          '</div>';
-        var ce = document.getElementById('sg-cancel');
-        if (ce) ce.addEventListener('click', function () { cancelled = true; clearInterval(tickTimer); clearTimeout(pollTimer); closeModal(); });
-      }
-
-      // Initial progress display.
-      showProgress(PHASES[0][1]);
-
-      // Timer tick updates the elapsed time + phase messages.
-      tickTimer = setInterval(function () {
-        var s = Math.round((Date.now() - t0) / 1000);
-        var te = document.getElementById('sg-timer'); if (te) te.textContent = s + 's';
-        var be = content.querySelector('[style*="background:var(--color-primary)"]'); if (be) be.style.width = Math.min(s/120*100, 98) + '%';
-        var pe = document.getElementById('sg-phase'); if (pe) {
-          var m = PHASES[0][1]; for (var i = 0; i < PHASES.length; i++) { if (s >= PHASES[i][0]) m = PHASES[i][1]; }
-          pe.textContent = m;
-        }
-      }, 1000);
-
-      // Start the analysis with optional focus text.
-      fetch('/agents/' + encodeURIComponent(agentId) + '/analyze', {
-        method: 'POST', credentials: 'same-origin',
-        headers: { 'Content-Type': 'application/x-www-form-urlencoded' },
-        body: focusText ? 'focus=' + encodeURIComponent(focusText) : '',
-      })
-      .then(function (r) { return r.json(); })
-      .then(function (startData) {
-        if (!startData.ok || !startData.runId) {
-          clearInterval(tickTimer);
-          content.innerHTML =
-            '<h3 style="margin:0 0 var(--space-3);">Analysis failed</h3>' +
-            '<div class="flash flash--error">' + esc(startData.error || 'Failed to start') + '</div>' +
-            '<div style="margin-top:var(--space-3);text-align:right;"><button type="button" class="btn btn--ghost btn--sm" data-close-modal="1">Close</button></div>';
-          return;
-        }
-
-        var runId = startData.runId;
-        var savedCurrentYaml = startData.currentYaml;
-
-        // Poll for progress + results.
-        function poll() {
-          if (cancelled) return;
-          fetch('/agents/' + encodeURIComponent(agentId) + '/analyze/' + encodeURIComponent(runId), { credentials: 'same-origin' })
-            .then(function (r) { return r.json(); })
-            .then(function (data) {
-              if (cancelled) return;
-              if (data.status === 'running') {
-                // Update phase from server-provided phase message or progress events.
-                var pe = document.getElementById('sg-phase');
-                if (pe) {
-                  if (data.phase) pe.textContent = data.phase;
-                  else if (data.progress && data.progress.length > 0) {
-                    var latest = data.progress[data.progress.length - 1];
-                    if (latest.message) pe.textContent = latest.message;
-                  }
-                }
-                pollTimer = setTimeout(poll, 2000);
-              } else if (data.status === 'done') {
-                clearInterval(tickTimer);
-                data.currentYaml = data.currentYaml || savedCurrentYaml;
-                renderResult(data);
-              } else {
-                clearInterval(tickTimer);
-                content.innerHTML =
-                  '<h3 style="margin:0 0 var(--space-3);">Analysis failed</h3>' +
-                  '<div class="flash flash--error">' + esc(data.error || 'Unknown error') + '</div>' +
-                  '<div style="margin-top:var(--space-3);text-align:right;"><button type="button" class="btn btn--ghost btn--sm" data-close-modal="1">Close</button></div>';
-              }
-            })
-            .catch(function () { if (!cancelled) pollTimer = setTimeout(poll, 3000); });
-        }
-
-        pollTimer = setTimeout(poll, 2000);
-      })
-      .catch(function (err) {
-        clearInterval(tickTimer);
-        content.innerHTML =
-          '<h3 style="margin:0 0 var(--space-3);">Error</h3>' +
-          '<div class="flash flash--error">' + esc(String(err)) + '</div>' +
-          '<div style="margin-top:var(--space-3);text-align:right;"><button type="button" class="btn btn--ghost btn--sm" data-close-modal="1">Close</button></div>';
-      });
-    }); // end sg-prompt-go click
-    }); // end suggest-btn click
-
-    modal.addEventListener('click', function (e) {
-      if (e.target === modal) closeModal();
-      if (e.target && e.target.getAttribute && e.target.getAttribute('data-close-modal')) closeModal();
-    });
-
-    // Auto-open from run detail page: ?suggest=1&focus=<error text>
-    var params = new URLSearchParams(window.location.search);
-    if (params.get('suggest') === '1') {
-      btn.click();
-      // Pre-fill focus after the modal renders.
-      setTimeout(function () {
-        var ta = document.getElementById('sg-focus');
-        var focusParam = params.get('focus');
-        if (ta && focusParam) ta.value = focusParam;
-      }, 50);
-    }
-  })();
-
-  // ── Pulse layout: containers, drag-and-drop, palette ────────────────
-  (function () {
-    var host = document.getElementById('pulse-containers');
-    var dataEl = document.getElementById('pulse-tile-data');
-    if (!host || !dataEl) return;
-
-    var tileData = JSON.parse(dataEl.textContent || '{}');
-    var allIds = tileData.allTileIds || [];
-    var systemIds = tileData.systemTileIds || [];
-
-    var LAYOUT_KEY = 'sua-pulse-layout';
-    var PALETTE_KEY = 'sua-pulse-palettes';
-    var PALETTES = ['default', 'dark', 'light', 'accent-teal', 'accent-red', 'accent-green'];
-
-    function getLayout() {
-      try { return JSON.parse(localStorage.getItem(LAYOUT_KEY) || 'null'); } catch { return null; }
-    }
-    function saveLayout(layout) {
-      try { localStorage.setItem(LAYOUT_KEY, JSON.stringify(layout)); } catch {}
-    }
-    function getPalettes() {
-      try { return JSON.parse(localStorage.getItem(PALETTE_KEY) || '{}'); } catch { return {}; }
-    }
-    function savePalettes(p) {
-      try { localStorage.setItem(PALETTE_KEY, JSON.stringify(p)); } catch {}
-    }
-
-    // Build default layout if none exists.
-    var layout = getLayout();
-    if (!layout || !layout.containers) {
-      layout = {
-        containers: [
-          { id: 'health', label: 'Health', tiles: systemIds.slice() },
-          { id: 'agents', label: 'Agents', tiles: allIds.filter(function(id) { return systemIds.indexOf(id) === -1; }) }
-        ]
-      };
-      saveLayout(layout);
-    }
-
-    // Merge any new tiles not in the layout into an "Other" container.
-    var assigned = {};
-    for (var ci = 0; ci < layout.containers.length; ci++) {
-      for (var ti = 0; ti < layout.containers[ci].tiles.length; ti++) {
-        assigned[layout.containers[ci].tiles[ti]] = true;
-      }
-    }
-    var unassigned = allIds.filter(function(id) { return !assigned[id]; });
-    if (unassigned.length > 0) {
-      var other = layout.containers.find(function(c) { return c.id === '_other'; });
-      if (!other) {
-        other = { id: '_other', label: 'Other', tiles: [] };
-        layout.containers.push(other);
-      }
-      for (var u = 0; u < unassigned.length; u++) other.tiles.push(unassigned[u]);
-      saveLayout(layout);
-    }
-
-    // Get all rendered tiles by agent ID.
-    var tileEls = {};
-    var allTileNodes = host.querySelectorAll('.pulse-tile[data-agent-id]');
-    for (var i = 0; i < allTileNodes.length; i++) {
-      tileEls[allTileNodes[i].getAttribute('data-agent-id')] = allTileNodes[i];
-    }
-
-    // Apply palettes: auto-palette as default, manual overrides on top.
-    var palettes = getPalettes();
-    for (var aid in tileEls) {
-      var manual = palettes[aid];
-      if (manual && manual !== 'default') {
-        tileEls[aid].setAttribute('data-palette', manual);
-      } else if (!manual) {
-        var auto = tileEls[aid].getAttribute('data-auto-palette');
-        if (auto) tileEls[aid].setAttribute('data-palette', auto);
-      }
-    }
-
-    // Re-render containers from layout.
-    function renderLayout() {
-      host.innerHTML = '';
-      for (var ci = 0; ci < layout.containers.length; ci++) {
-        var c = layout.containers[ci];
-        var section = document.createElement('section');
-        section.className = 'pulse-container';
-        section.setAttribute('data-container-id', c.id);
-
-        // Header.
-        var header = document.createElement('div');
-        header.className = 'pulse-container__header';
-        var label = document.createElement('span');
-        label.className = 'pulse-container__label';
-        label.contentEditable = 'true';
-        label.textContent = c.label;
-        label.addEventListener('blur', (function(cont) { return function(e) {
-          cont.label = e.target.textContent.trim() || cont.label;
-          saveLayout(layout);
-        }; })(c));
-        label.addEventListener('keydown', function(e) {
-          if (e.key === 'Enter') { e.preventDefault(); e.target.blur(); }
-        });
-        header.appendChild(label);
-
-        var delBtn = document.createElement('button');
-        delBtn.className = 'pulse-container__delete';
-        delBtn.title = 'Remove group';
-        delBtn.textContent = '\u00D7';
-        delBtn.addEventListener('click', (function(cont) { return function() {
-          // Move tiles to Other.
-          var otherCont = layout.containers.find(function(x) { return x.id === '_other'; });
-          if (!otherCont) {
-            otherCont = { id: '_other', label: 'Other', tiles: [] };
-            layout.containers.push(otherCont);
-          }
-          for (var t = 0; t < cont.tiles.length; t++) otherCont.tiles.push(cont.tiles[t]);
-          layout.containers = layout.containers.filter(function(x) { return x.id !== cont.id; });
-          saveLayout(layout);
-          renderLayout();
-          restoreCollapsed();
-          restorePalettes();
-          applySizes();
-          setEditMode(editMode);
-        }; })(c));
-        header.appendChild(delBtn);
-        section.appendChild(header);
-
-        // Grid.
-        var grid = document.createElement('div');
-        grid.className = 'pulse-grid';
-        grid.setAttribute('data-container-id', c.id);
-
-        var hasTiles = false;
-        for (var ti = 0; ti < c.tiles.length; ti++) {
-          var el = tileEls[c.tiles[ti]];
-          if (el) { grid.appendChild(el); hasTiles = true; }
-        }
-
-        if (!hasTiles) {
-          var empty = document.createElement('div');
-          empty.className = 'pulse-container__empty';
-          empty.textContent = 'Drag widgets here';
-          grid.appendChild(empty);
-        }
-
-        section.appendChild(grid);
-        host.appendChild(section);
-      }
-    }
-
-    renderLayout();
-
-    // ── Edit mode toggle ─────────────────────────────────────────────
-    var editMode = false;
-    var editBtn = document.getElementById('pulse-edit-toggle');
-    var addContainerBtn = document.getElementById('pulse-add-container');
-
-    function setEditMode(on) {
-      editMode = on;
-      document.body.classList.toggle('pulse-edit-mode', on);
-      if (editBtn) editBtn.textContent = on ? '\u2713 Done editing' : '\u270E Edit layout';
-      if (editBtn) editBtn.classList.toggle('btn--primary', on);
-      if (editBtn) editBtn.classList.toggle('btn--ghost', !on);
-      if (addContainerBtn) addContainerBtn.style.display = on ? '' : 'none';
-      // Toggle draggable on all tiles.
-      var allTiles = document.querySelectorAll('.pulse-tile[data-agent-id]');
-      for (var i = 0; i < allTiles.length; i++) {
-        allTiles[i].setAttribute('draggable', on ? 'true' : 'false');
-      }
-    }
-
-    if (editBtn) {
-      editBtn.addEventListener('click', function() { setEditMode(!editMode); });
-    }
-
-    // ── Drag and drop ────────────────────────────────────────────────
-    var draggedId = null;
-    var draggedEl = null;
-
-    document.addEventListener('dragstart', function(e) {
-      if (!editMode) { e.preventDefault(); return; }
-      var tile = e.target.closest ? e.target.closest('.pulse-tile[data-agent-id]') : null;
-      if (!tile) return;
-      draggedId = tile.getAttribute('data-agent-id');
-      draggedEl = tile;
-      tile.classList.add('pulse-tile--dragging');
-      e.dataTransfer.effectAllowed = 'move';
-      e.dataTransfer.setData('text/plain', draggedId);
-    });
-
-    document.addEventListener('dragend', function() {
-      if (draggedEl) draggedEl.classList.remove('pulse-tile--dragging');
-      draggedId = null;
-      draggedEl = null;
-      // Remove all drop targets.
-      var targets = document.querySelectorAll('.pulse-container--drop-target');
-      for (var i = 0; i < targets.length; i++) targets[i].classList.remove('pulse-container--drop-target');
-    });
-
-    document.addEventListener('dragover', function(e) {
-      var grid = e.target.closest ? e.target.closest('.pulse-grid') : null;
-      if (!grid || !draggedId) return;
-      e.preventDefault();
-      e.dataTransfer.dropEffect = 'move';
-      var container = grid.closest('.pulse-container');
-      if (container) container.classList.add('pulse-container--drop-target');
-    });
-
-    document.addEventListener('dragleave', function(e) {
-      var container = e.target.closest ? e.target.closest('.pulse-container') : null;
-      if (container && !container.contains(e.relatedTarget)) {
-        container.classList.remove('pulse-container--drop-target');
-      }
-    });
-
-    document.addEventListener('drop', function(e) {
-      var grid = e.target.closest ? e.target.closest('.pulse-grid') : null;
-      if (!grid || !draggedId) return;
-      e.preventDefault();
-
-      var targetContainerId = grid.getAttribute('data-container-id');
-      if (!targetContainerId) return;
-
-      // Find drop position: which tile are we dropping before?
-      var tiles = grid.querySelectorAll('.pulse-tile[data-agent-id]');
-      var dropIndex = tiles.length; // default: end
-      for (var i = 0; i < tiles.length; i++) {
-        var rect = tiles[i].getBoundingClientRect();
-        var midY = rect.top + rect.height / 2;
-        var midX = rect.left + rect.width / 2;
-        if (e.clientY < midY || (e.clientY < rect.bottom && e.clientX < midX)) {
-          dropIndex = i;
-          break;
-        }
-      }
-
-      // Remove from source container in layout.
-      for (var ci = 0; ci < layout.containers.length; ci++) {
-        var idx = layout.containers[ci].tiles.indexOf(draggedId);
-        if (idx !== -1) {
-          layout.containers[ci].tiles.splice(idx, 1);
-          break;
-        }
-      }
-
-      // Insert into target container.
-      var targetCont = layout.containers.find(function(c) { return c.id === targetContainerId; });
-      if (targetCont) {
-        // Adjust drop index if the dragged element was before the drop point in the same container.
-        targetCont.tiles.splice(dropIndex, 0, draggedId);
-      }
-
-      saveLayout(layout);
-      renderLayout();
-      restoreCollapsed();
-      restorePalettes();
-      applySizes();
-      setEditMode(editMode);
-    });
-
-    // ── Add container ────────────────────────────────────────────────
-    var addBtn = document.getElementById('pulse-add-container');
-    if (addBtn) {
-      addBtn.addEventListener('click', function() {
-        var name = prompt('Group name:');
-        if (!name || !name.trim()) return;
-        var id = 'custom-' + Date.now();
-        layout.containers.push({ id: id, label: name.trim(), tiles: [] });
-        saveLayout(layout);
-        renderLayout();
-        restoreCollapsed();
-        restorePalettes();
-        applySizes();
-        setEditMode(editMode);
-      });
-    }
-
-    // ── Palette cycling ──────────────────────────────────────────────
-    document.addEventListener('click', function(e) {
-      var btn = e.target.closest ? e.target.closest('.pulse-tile__palette-btn') : null;
-      if (!btn) return;
-      var tileId = btn.getAttribute('data-tile-id');
-      if (!tileId) return;
-      var tile = document.querySelector('.pulse-tile[data-agent-id=\"' + tileId + '\"]');
-      if (!tile) return;
-
-      var current = tile.getAttribute('data-palette') || 'default';
-      var idx = PALETTES.indexOf(current);
-      var next = PALETTES[(idx + 1) % PALETTES.length];
-
-      if (next === 'default') {
-        tile.removeAttribute('data-palette');
-      } else {
-        tile.setAttribute('data-palette', next);
-      }
-
-      var p = getPalettes();
-      if (next === 'default') { delete p[tileId]; } else { p[tileId] = next; }
-      savePalettes(p);
-    });
-
-    function restorePalettes() {
-      var p = getPalettes();
-      // Apply auto-palettes first, then manual overrides.
-      var allT = document.querySelectorAll('.pulse-tile[data-agent-id]');
-      for (var i = 0; i < allT.length; i++) {
-        var id = allT[i].getAttribute('data-agent-id');
-        var manual = p[id];
-        if (manual && manual !== 'default') {
-          allT[i].setAttribute('data-palette', manual);
-        } else if (!manual) {
-          // Use auto-palette from server-rendered attribute.
-          var auto = allT[i].getAttribute('data-auto-palette');
-          if (auto) allT[i].setAttribute('data-palette', auto);
-        }
-      }
-    }
-
-    // ── Tile resize via drag handle ─────────────────────────────────
-    // Mousedown on the bottom-right handle, track mouse, snap to grid
-    // columns (1-4) and rows (1-4), update inline style + localStorage.
-    var SIZES_KEY = 'sua-pulse-sizes';
-    function getSizes() {
-      try { return JSON.parse(localStorage.getItem(SIZES_KEY) || '{}'); } catch { return {}; }
-    }
-    function saveSizes(s) {
-      try { localStorage.setItem(SIZES_KEY, JSON.stringify(s)); } catch {}
-    }
-
-    function applySizes() {
-      var sizes = getSizes();
-      var allT = document.querySelectorAll('.pulse-tile[data-agent-id]');
-      for (var i = 0; i < allT.length; i++) {
-        var id = allT[i].getAttribute('data-agent-id');
-        var s = sizes[id];
-        if (s) {
-          allT[i].style.gridColumn = 'span ' + (s.cols || 1);
-          allT[i].style.gridRow = 'span ' + (s.rows || 1);
-          // Remove preset size classes — inline style takes over.
-          allT[i].classList.remove('pulse-tile--2x1', 'pulse-tile--1x2', 'pulse-tile--2x2');
-        }
-      }
-    }
-
-    var resizing = null; // { tile, startX, startY, startCols, startRows, gridCellW, gridCellH, gap }
-
-    document.addEventListener('mousedown', function(e) {
-      if (!editMode) return;
-      var handle = e.target.closest ? e.target.closest('.pulse-tile__resize-handle') : null;
-      if (!handle) return;
-      e.preventDefault();
-      e.stopPropagation();
-
-      var tile = handle.closest('.pulse-tile');
-      if (!tile) return;
-
-      var grid = tile.closest('.pulse-grid');
-      if (!grid) return;
-
-      // Compute grid cell dimensions from the grid's actual layout.
-      var gridRect = grid.getBoundingClientRect();
-      var gridStyle = window.getComputedStyle(grid);
-      var gapStr = gridStyle.gap || gridStyle.gridGap || '16px';
-      var gap = parseInt(gapStr) || 16;
-      var cols = 4; // fixed 4-column grid
-      var gridCellW = (gridRect.width - gap * (cols - 1)) / cols;
-      var gridCellH = gridCellW * 0.6; // approximate row height (aspect ratio)
-
-      var tileRect = tile.getBoundingClientRect();
-      var currentCols = Math.round(tileRect.width / (gridCellW + gap)) || 1;
-      var currentRows = Math.max(1, Math.round(tileRect.height / (gridCellH + gap)));
-
-      tile.classList.add('pulse-tile--resizing');
-
-      resizing = {
-        tile: tile,
-        agentId: tile.getAttribute('data-agent-id'),
-        startX: e.clientX,
-        startY: e.clientY,
-        startCols: currentCols,
-        startRows: currentRows,
-        gridCellW: gridCellW,
-        gridCellH: gridCellH,
-        gap: gap
-      };
-    });
-
-    document.addEventListener('mousemove', function(e) {
-      if (!resizing) return;
-      e.preventDefault();
-
-      var dx = e.clientX - resizing.startX;
-      var dy = e.clientY - resizing.startY;
-
-      var newCols = Math.max(1, Math.min(4, resizing.startCols + Math.round(dx / (resizing.gridCellW + resizing.gap))));
-      var newRows = Math.max(1, Math.min(4, resizing.startRows + Math.round(dy / (resizing.gridCellH + resizing.gap))));
-
-      resizing.tile.style.gridColumn = 'span ' + newCols;
-      resizing.tile.style.gridRow = 'span ' + newRows;
-      resizing.newCols = newCols;
-      resizing.newRows = newRows;
-    });
-
-    document.addEventListener('mouseup', function() {
-      if (!resizing) return;
-
-      resizing.tile.classList.remove('pulse-tile--resizing');
-      resizing.tile.classList.remove('pulse-tile--2x1', 'pulse-tile--1x2', 'pulse-tile--2x2');
-
-      var newCols = resizing.newCols || resizing.startCols;
-      var newRows = resizing.newRows || resizing.startRows;
-
-      // Persist size override.
-      var sizes = getSizes();
-      if (newCols === 1 && newRows === 1) {
-        delete sizes[resizing.agentId]; // back to default
-        resizing.tile.style.gridColumn = '';
-        resizing.tile.style.gridRow = '';
-      } else {
-        sizes[resizing.agentId] = { cols: newCols, rows: newRows };
-      }
-      saveSizes(sizes);
-      resizing = null;
-    });
-
-    applySizes();
-
-    function restoreCollapsed() {
-      // Re-apply collapsed state after layout re-render.
-      try {
-        var collapsed = JSON.parse(localStorage.getItem('sua-pulse-collapsed') || '{}');
-        for (var id in collapsed) {
-          var el = document.querySelector('.pulse-tile[data-agent-id=\"' + id + '\"]');
-          if (el) el.classList.add('pulse-tile--collapsed');
-        }
-      } catch {}
-    }
-    restoreCollapsed();
-  })();
-
-  // ── Pulse tile collapse/expand ──────────────────────────────────────
-  // Clicking the chevron or header title toggles the tile body.
-  // Collapsed state persists in localStorage per agent id.
-  (function () {
-    var STORAGE_KEY = 'sua-pulse-collapsed';
-    function getCollapsed() {
-      try { return JSON.parse(localStorage.getItem(STORAGE_KEY) || '{}'); } catch { return {}; }
-    }
-    function setCollapsed(map) {
-      try { localStorage.setItem(STORAGE_KEY, JSON.stringify(map)); } catch {}
-    }
-
-    // Restore collapsed state on load.
-    var collapsed = getCollapsed();
-    var tiles = document.querySelectorAll('.pulse-tile');
-    for (var i = 0; i < tiles.length; i++) {
-      var btn = tiles[i].querySelector('.pulse-tile__collapse');
-      if (btn) {
-        var id = btn.getAttribute('data-tile-id');
-        if (id && collapsed[id]) tiles[i].classList.add('pulse-tile--collapsed');
-      }
-    }
-
-    // Toggle on click.
-    document.addEventListener('click', function (e) {
-      var target = e.target;
-      // Check if click is on the collapse button or the header trigger area.
-      var tileId = null;
-      if (target.classList && target.classList.contains('pulse-tile__collapse')) {
-        tileId = target.getAttribute('data-tile-id');
-      } else if (target.closest && target.closest('[data-collapse-trigger]')) {
-        tileId = target.closest('[data-collapse-trigger]').getAttribute('data-tile-id');
-      }
-      if (!tileId) return;
-
-      // Find the parent tile.
-      var tile = target.closest('.pulse-tile');
-      if (!tile) return;
-
-      tile.classList.toggle('pulse-tile--collapsed');
-      var map = getCollapsed();
-      if (tile.classList.contains('pulse-tile--collapsed')) {
-        map[tileId] = true;
-      } else {
-        delete map[tileId];
-      }
-      setCollapsed(map);
-    });
-  })();
-
-  // ── Pulse media: click-to-play YouTube embeds ───────────────────────
-  // Click thumbnail to swap to iframe embed. If embed fails (video
-  // disallows embedding), show a "blocked" message with link.
-  (function () {
-    document.addEventListener('click', function (e) {
-      var el = e.target;
-      while (el && !el.classList?.contains('pulse-media-yt')) el = el.parentElement;
-      if (!el) return;
-      var embedUrl = el.getAttribute('data-embed');
-      var watchUrl = el.getAttribute('data-watch') || '';
-      if (!embedUrl) return;
-      e.preventDefault();
-
-      // Save the original thumbnail HTML so we can restore on error.
-      var originalHtml = el.innerHTML;
-      el.style.cursor = 'default';
-
-      var iframe = document.createElement('iframe');
-      iframe.src = embedUrl;
-      iframe.style.cssText = 'width:100%;aspect-ratio:16/9;border:0;border-radius:var(--radius-sm);';
-      iframe.allow = 'accelerometer;autoplay;clipboard-write;encrypted-media;gyroscope;picture-in-picture';
-      iframe.allowFullscreen = true;
-
-      // Detect embed failures. YouTube returns a 200 with an error page,
-      // so we can't catch HTTP errors. Instead, set a timeout: if the
-      // iframe loads the error page, the user can click "open externally".
-      el.innerHTML = '';
-      el.appendChild(iframe);
-
-      // Add a small "can't play? open externally" link below.
-      if (watchUrl) {
-        var fallback = document.createElement('div');
-        fallback.style.cssText = 'text-align:center;margin-top:var(--space-1);font-size:var(--font-size-xs);';
-        fallback.innerHTML = '<a href="' + watchUrl + '" target="_blank" rel="noopener" style="color:var(--color-text-muted);">Not loading? Open on YouTube \u2197</a>';
-        el.parentElement.appendChild(fallback);
-      }
     });
   })();
 
@@ -1515,173 +440,5 @@ export const DASHBOARD_JS = `
     });
   })();
 
-  // ── Build from goal wizard ─────────────────────────────────────────
-  // Modal wizard that designs a new agent from a prompt.
-  (function () {
-    var btn = document.getElementById('build-from-goal-btn');
-    var modal = document.getElementById('build-modal');
-    var content = document.getElementById('build-modal-content');
-    if (!btn || !modal || !content) return;
-
-    function esc(s) { var d = document.createElement('div'); d.textContent = s; return d.innerHTML; }
-    function closeModal() { modal.classList.remove('is-open'); }
-
-    btn.addEventListener('click', function () { modal.classList.add('is-open'); });
-
-    modal.addEventListener('click', function (e) {
-      if (e.target === modal) closeModal();
-      if (e.target && e.target.getAttribute && e.target.getAttribute('data-close-build')) closeModal();
-    });
-
-    var submitBtn = document.getElementById('build-submit-btn');
-    if (!submitBtn) return;
-
-    submitBtn.addEventListener('click', function () {
-      var goalEl = document.getElementById('build-goal');
-      var focusEl = document.getElementById('build-focus');
-      var goal = goalEl ? goalEl.value.trim() : '';
-      if (!goal) { goalEl && goalEl.focus(); return; }
-      var focus = focusEl ? focusEl.value.trim() : '';
-
-      var t0 = Date.now();
-      var cancelled = false;
-      var pollTimer = null;
-      var PHASES = [[0,'Designing agent...'],[5,'Selecting tools...'],[15,'Building node pipeline...'],[30,'Generating YAML...'],[60,'Still working...'],[90,'Almost there...']];
-
-      content.innerHTML =
-        '<div style="padding:var(--space-4);">' +
-        '<div style="display:flex;align-items:center;gap:var(--space-3);margin-bottom:var(--space-4);">' +
-          '<div class="spinner"></div>' +
-          '<div style="flex:1;"><div style="font-weight:var(--weight-medium);">Building agent</div>' +
-          '<div class="dim" style="font-size:var(--font-size-xs);" id="build-phase">Designing agent...</div></div>' +
-          '<div style="font-family:var(--font-mono);font-size:var(--font-size-lg);color:var(--color-text-muted);min-width:3rem;text-align:right;" id="build-timer">0s</div>' +
-        '</div>' +
-        '<div style="height:4px;background:var(--color-border);border-radius:2px;overflow:hidden;margin-bottom:var(--space-3);">' +
-          '<div id="build-bar" style="height:100%;background:var(--color-primary);border-radius:2px;width:0%;transition:width 1s linear;"></div>' +
-        '</div>' +
-        '<div style="text-align:right;"><button type="button" class="btn btn--ghost btn--sm" id="build-cancel">Cancel</button></div></div>';
-
-      var serverPhase = '';
-      var tickTimer = setInterval(function () {
-        var s = Math.round((Date.now() - t0) / 1000);
-        var te = document.getElementById('build-timer'); if (te) te.textContent = s + 's';
-        var be = document.getElementById('build-bar'); if (be) be.style.width = Math.min(s/120*100, 98) + '%';
-        // Only use timer-based fallback phases when no server phase has arrived.
-        if (!serverPhase) {
-          var pe = document.getElementById('build-phase'); if (pe) {
-            var m = PHASES[0][1]; for (var i = 0; i < PHASES.length; i++) { if (s >= PHASES[i][0]) m = PHASES[i][1]; }
-            pe.textContent = m;
-          }
-        }
-      }, 1000);
-
-      var ce = document.getElementById('build-cancel');
-      if (ce) ce.addEventListener('click', function () { cancelled = true; clearInterval(tickTimer); clearTimeout(pollTimer); closeModal(); });
-
-      fetch('/agents/build', {
-        method: 'POST', credentials: 'same-origin',
-        headers: { 'Content-Type': 'application/x-www-form-urlencoded' },
-        body: 'goal=' + encodeURIComponent(goal) + (focus ? '&focus=' + encodeURIComponent(focus) : ''),
-      })
-      .then(function (r) { return r.json(); })
-      .then(function (startData) {
-        if (!startData.ok || !startData.runId) {
-          clearInterval(tickTimer);
-          content.innerHTML = '<div class="flash flash--error">' + esc(startData.error || 'Failed') + '</div>' +
-            '<div style="margin-top:var(--space-3);text-align:right;"><button type="button" class="btn btn--ghost btn--sm" data-close-build="1">Close</button></div>';
-          return;
-        }
-        var runId = startData.runId;
-
-        function poll() {
-          if (cancelled) return;
-          fetch('/agents/build/' + encodeURIComponent(runId), { credentials: 'same-origin' })
-            .then(function (r) { return r.json(); })
-            .then(function (data) {
-              if (cancelled) return;
-              if (data.status === 'running') {
-                if (data.phase) {
-                  serverPhase = data.phase;
-                  var pe = document.getElementById('build-phase');
-                  if (pe) pe.textContent = data.phase;
-                }
-                pollTimer = setTimeout(poll, 2000);
-              } else if (data.status === 'done') {
-                clearInterval(tickTimer);
-                var h = '<h3 style="margin:0 0 var(--space-3);">Agent designed</h3>';
-                if (data.agentName) h += '<p style="font-weight:var(--weight-medium);margin:0 0 var(--space-2);">' + esc(data.agentName) + ' <span class="dim">(' + esc(data.agentId) + ')</span></p>';
-                if (data.yamlError) {
-                  h += '<div class="flash flash--error" style="margin-bottom:var(--space-3);font-size:var(--font-size-xs);">' +
-                    '<strong>Validation error:</strong> ' + esc(data.yamlError) +
-                    '<br>Edit the YAML below to fix, then create.</div>';
-                }
-                if (data.yaml) {
-                  h += '<label style="display:flex;flex-direction:column;gap:var(--space-1);margin-bottom:var(--space-3);">' +
-                    '<span class="dim" style="font-size:var(--font-size-xs);font-weight:var(--weight-semibold);">Review and edit YAML before creating</span>' +
-                    '<textarea id="build-yaml-editor" rows="15" ' +
-                    'style="padding:var(--space-3);border:1px solid var(--color-border-strong);border-radius:var(--radius-sm);' +
-                    'font-family:var(--font-mono);font-size:var(--font-size-xs);resize:vertical;line-height:1.5;tab-size:2;">' +
-                    esc(data.yaml) + '</textarea></label>';
-                }
-                h += '<div id="build-result-flash"></div>';
-                h += '<div style="display:flex;gap:var(--space-2);justify-content:flex-end;flex-wrap:wrap;">';
-                if (data.yaml) {
-                  h += '<button type="button" class="btn btn--primary btn--sm" id="build-create-btn">Create agent</button>';
-                }
-                h += '<button type="button" class="btn btn--ghost btn--sm" data-close-build="1">Dismiss</button></div>';
-                content.innerHTML = h;
-
-                var createBtn = document.getElementById('build-create-btn');
-                if (createBtn) {
-                  createBtn.addEventListener('click', function () {
-                    var yamlEditor = document.getElementById('build-yaml-editor');
-                    var yamlText = yamlEditor ? yamlEditor.value : '';
-                    if (!yamlText.trim()) return;
-                    createBtn.disabled = true;
-                    createBtn.textContent = 'Creating...';
-                    var flashEl = document.getElementById('build-result-flash');
-                    if (flashEl) flashEl.innerHTML = '';
-                    fetch('/agents/build/create', {
-                      method: 'POST', credentials: 'same-origin',
-                      headers: { 'Content-Type': 'application/json' },
-                      body: JSON.stringify({ yaml: yamlText }),
-                    })
-                    .then(function (r) { return r.json(); })
-                    .then(function (result) {
-                      if (result.ok) {
-                        window.location.href = '/agents/' + encodeURIComponent(result.agentId);
-                      } else {
-                        createBtn.disabled = false;
-                        createBtn.textContent = 'Create agent';
-                        // If agent already exists, show a link to it.
-                        var msg = result.error || 'Creation failed';
-                        var existsMatch = msg.match(/["']([a-z0-9][a-z0-9-]*)["'] already exists/i);
-                        if (existsMatch && flashEl) {
-                          flashEl.innerHTML = '<div class="flash flash--error" style="margin-bottom:var(--space-3);font-size:var(--font-size-xs);">' +
-                            esc(msg) + ' <a href="/agents/' + encodeURIComponent(existsMatch[1]) + '" style="font-weight:var(--weight-semibold);">Open existing agent \u2192</a></div>';
-                        } else if (flashEl) {
-                          flashEl.innerHTML = '<div class="flash flash--error" style="margin-bottom:var(--space-3);font-size:var(--font-size-xs);">' + esc(msg) + '</div>';
-                        }
-                      }
-                    });
-                  });
-                }
-              } else {
-                clearInterval(tickTimer);
-                content.innerHTML = '<div class="flash flash--error">' + esc(data.error || 'Build failed') + '</div>' +
-                  '<div style="margin-top:var(--space-3);text-align:right;"><button type="button" class="btn btn--ghost btn--sm" data-close-build="1">Close</button></div>';
-              }
-            })
-            .catch(function () { if (!cancelled) pollTimer = setTimeout(poll, 3000); });
-        }
-        pollTimer = setTimeout(poll, 2000);
-      })
-      .catch(function (err) {
-        clearInterval(tickTimer);
-        content.innerHTML = '<div class="flash flash--error">' + esc(String(err)) + '</div>' +
-          '<div style="margin-top:var(--space-3);text-align:right;"><button type="button" class="btn btn--ghost btn--sm" data-close-build="1">Close</button></div>';
-      });
-    });
-  })();
 })();
 `;
